refactor folder structure
This commit establishes the folder structure for cstorpool
auto project:
- lib/ has the core business logic
- monitor/ to have monitoring logic
- e2e/ to have e2e test logic
- test/integration to have integration test logic
- manifests/ to have deployment related documents
