+++ lib/api.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cspauto
+package lib
 
 import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
