+++ lib/sync-cspcauto.go
@@ -14,9 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cspauto
+package lib
 
-// Sync reconciles cspauto request
-func Sync(req *GenericHookRequest) (*GenericHookResponse, error) {
+// SyncCSPCAuto reconciles cspcauto request
+func SyncCSPCAuto(req *GenericHookRequest) (*GenericHookResponse, error) {
 	return &GenericHookResponse{}, nil
 }
