+++ manifests/operator.yaml
@@ -1,46 +1,46 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: cstorpoolautos.dao.mayadata.io
+  name: cspcautos.dao.mayadata.io
 spec:
   group: dao.mayadata.io
   version: v1alpha1
   scope: Namespaced
   names:
-    plural: cstorpoolautos
-    singular: cstorpoolauto
-    kind: CStorPoolAuto
+    plural: cspcautos
+    singular: cspcauto
+    kind: CSPCAuto
     shortNames:
-    - cspauto
+    - cspcauto
 ---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: cspauto
+  name: cspcauto
   namespace: dao
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: cspauto
+      app: cspcauto
   template:
     metadata:
       labels:
-        app: cspauto
+        app: cspcauto
     spec:
       containers:
-      - name: controller
-        image: quay.io/amitkumardas/cspauto:latest
-        command: ["cspauto"]
+      - name: server
+        image: quay.io/amitkumardas/cspcauto:latest
+        command: ["cspcauto"]
 ---
 apiVersion: v1
 kind: Service
 metadata:
-  name: cspauto
+  name: cspcauto
   namespace: dao
 spec:
   selector:
-    app: csp-auto
+    app: cspcauto
   ports:
   - port: 80
     targetPort: 8080
